clean up CityBMWList query renderer
