+++ src/lib/Containers/index.tsx
@@ -1,5 +1,4 @@
 import { ArtworkContainer as Artwork } from "lib/Scenes/Artwork"
-import CityBMWList from "lib/Scenes/City/CityBMWList"
 import CityFairList from "lib/Scenes/City/CityFairList"
 import CitySavedList from "lib/Scenes/City/CitySavedList"
 import CitySectionList from "lib/Scenes/City/CitySectionList"
@@ -12,7 +11,6 @@ import Show from "../Scenes/Show"
 export default {
   Artwork,
   CitySavedList,
-  CityBMWList,
   CitySectionList,
   CityFairList,
   Fair,
