--- src/lib/Scenes/City/__tests__/CityBMWList-tests.tsx
@@ -1,5 +1,0 @@
-import CityBMWList from "../CityBMWList"
-
-describe(CityBMWList, () => {
-  it.todo("No tests yet, it's just placeholders so far.")
-})
